[CALCITE-2345] Running Unit tests with Fongo and integration tests with real mongo instance (Andrei Sereda)
Better test coverage for unit tests using Fongo in-memory implementation of Mongo API.
New code will decide (at runtime) what connection to make: fongo vs mongo. Identical tests will be run against
both databases (depending on maven profile surefire vs failsafe)
