+++ mongodb/src/test/java/org/apache/calcite/adapter/mongodb/MongoAdapterTest.java
@@ -16,143 +16,749 @@
  */
 package org.apache.calcite.adapter.mongodb;
 
-import org.apache.calcite.jdbc.CalciteConnection;
+import org.apache.calcite.schema.Schema;
+import org.apache.calcite.schema.SchemaFactory;
 import org.apache.calcite.schema.SchemaPlus;
 import org.apache.calcite.test.CalciteAssert;
+import org.apache.calcite.test.MongoAssertions;
 
-import com.github.fakemongo.junit.FongoRule;
+
+import org.apache.calcite.util.Bug;
+import org.apache.calcite.util.Util;
+
+import com.google.common.base.Function;
+import com.google.common.base.Preconditions;
+import com.google.common.io.LineProcessor;
+import com.google.common.io.Resources;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 
+import net.hydromatic.foodmart.data.json.FoodmartJson;
+
+import org.bson.BsonDateTime;
 import org.bson.BsonDocument;
+import org.bson.BsonInt32;
+import org.bson.BsonString;
 import org.bson.Document;
 
-import org.junit.Before;
+import org.hamcrest.CoreMatchers;
+
+import org.junit.Assert;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.Ignore;
-import org.junit.Rule;
 import org.junit.Test;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
 
 /**
- * Tests current adapter using in-memory (fake) implementation of Mongo API:
- * <a href="https://github.com/fakemongo/fongo">Fongo</a>.
- *
+ * Testing mongo adapter functionality. By default runs with
+ * <a href="https://github.com/fakemongo/fongo">Fongo</a> unless {@code IT} maven profile is enabled
+ * (via {@code $ mvn -Pit install}).
  */
-public class MongoAdapterTest {
+public class MongoAdapterTest implements SchemaFactory {
 
-  @Rule
-  public final FongoRule rule = new FongoRule();
+  /** Connection factory based on the "mongo-zips" model. */
+  protected static final URL MODEL = MongoAdapterTest.class.getResource("/mongo-model.json");
 
-  private MongoDatabase mongoDb;
-  private MongoCollection<Document> zips;
+  /** Number of records in local file */
+  protected static final int ZIPS_SIZE = 149;
 
-  @Before
-  public void setUp() throws Exception {
-    mongoDb = rule.getDatabase(getClass().getSimpleName());
-    zips = mongoDb.getCollection("zips");
+  @ClassRule
+  public static final MongoDatabaseRule RULE = MongoDatabaseRule.create();
+
+  private static MongoSchema schema;
+
+  @BeforeClass
+  public static void setUp() throws Exception {
+    MongoDatabase database = RULE.database();
+
+    populate(database.getCollection("zips"), MongoAdapterTest.class.getResource("/zips-mini.json"));
+    populate(database.getCollection("store"), FoodmartJson.class.getResource("/store.json"));
+    populate(database.getCollection("warehouse"),
+            FoodmartJson.class.getResource("/warehouse.json"));
+
+    // Manually insert data for data-time test.
+    MongoCollection<BsonDocument> datatypes =  database.getCollection("datatypes")
+            .withDocumentClass(BsonDocument.class);
+    if (datatypes.count() > 0) {
+      datatypes.deleteMany(new BsonDocument());
+    }
+    BsonDocument doc = new BsonDocument();
+    Date date = new SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).parse("2012-09-05");
+    doc.put("date", new BsonDateTime(date.getTime()));
+    doc.put("value", new BsonInt32(1231));
+    doc.put("ownerId", new BsonString("531e7789e4b0853ddb861313"));
+    datatypes.insertOne(doc);
+
+    schema = new MongoSchema(database);
+  }
+
+  private static void populate(MongoCollection<Document> collection, URL resource)
+          throws IOException {
+    Preconditions.checkNotNull(collection, "collection");
+
+    if (collection.count() > 0) {
+      // delete any existing documents (run from a clean set)
+      collection.deleteMany(new BsonDocument());
+    }
+
+    MongoCollection<BsonDocument> bsonCollection = collection.withDocumentClass(BsonDocument.class);
+    Resources.readLines(resource, StandardCharsets.UTF_8, new LineProcessor<Void>() {
+      @Override public boolean processLine(String line) throws IOException {
+        bsonCollection.insertOne(BsonDocument.parse(line));
+        return true;
+      }
+
+      @Override public Void getResult() {
+        return null;
+      }
+    });
   }
 
   /**
-   * Handcrafted connection where we manually added {@link MongoSchema}
+   *  Returns always the same schema to avoid initialization costs.
    */
-  private CalciteAssert.ConnectionFactory newConnectionFactory() {
-    return new CalciteAssert.ConnectionFactory() {
-      @Override public Connection createConnection() throws SQLException {
-        Connection connection = DriverManager.getConnection("jdbc:calcite:");
-        final SchemaPlus root = connection.unwrap(CalciteConnection.class).getRootSchema();
-        root.add("mongo", new MongoSchema(mongoDb));
-        return connection;
+  @Override public Schema create(SchemaPlus parentSchema, String name,
+                                 Map<String, Object> operand) {
+    return schema;
+  }
+
+  private CalciteAssert.AssertThat assertModel(String model) {
+    // ensure that Schema from this instance is being used
+    model = model.replace(MongoSchemaFactory.class.getName(), MongoAdapterTest.class.getName());
+
+    return CalciteAssert.that()
+            .withModel(model);
+  }
+
+  private CalciteAssert.AssertThat assertModel(URL url) {
+    Preconditions.checkNotNull(url, "url");
+    try {
+      return assertModel(Resources.toString(url, StandardCharsets.UTF_8));
+    } catch (IOException e) {
+      throw new UncheckedIOException(e);
+    }
+  }
+
+  @Test
+  public void testSort() {
+    assertModel(MODEL)
+            .query("select * from zips order by state")
+            .returnsCount(ZIPS_SIZE)
+            .explainContains("PLAN=MongoToEnumerableConverter\n"
+                    + "  MongoSort(sort0=[$4], dir0=[ASC])\n"
+                    + "    MongoProject(CITY=[CAST(ITEM($0, 'city')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"], LONGITUDE=[CAST(ITEM(ITEM($0, 'loc'), 0)):FLOAT], LATITUDE=[CAST(ITEM(ITEM($0, 'loc'), 1)):FLOAT], POP=[CAST(ITEM($0, 'pop')):INTEGER], STATE=[CAST(ITEM($0, 'state')):VARCHAR(2) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"], ID=[CAST(ITEM($0, '_id')):VARCHAR(5) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"])\n"
+                    + "      MongoTableScan(table=[[mongo_raw, zips]])");
+  }
+
+  @Test public void testSortLimit() {
+    assertModel(MODEL)
+            .query("select state, id from zips\n"
+                    + "order by state, id offset 2 rows fetch next 3 rows only")
+            .returns("STATE=AK; ID=99801\n"
+                    + "STATE=AL; ID=35215\n"
+                    + "STATE=AL; ID=35401\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state', ID: '$_id'}}",
+                            "{$sort: {STATE: 1, ID: 1}}",
+                            "{$skip: 2}",
+                            "{$limit: 3}"));
+  }
+
+  @Test public void testOffsetLimit() {
+    assertModel(MODEL)
+            .query("select state, id from zips\n"
+                    + "offset 2 fetch next 3 rows only")
+            .runs()
+            .queryContains(
+                    mongoChecker(
+                            "{$skip: 2}",
+                            "{$limit: 3}",
+                            "{$project: {STATE: '$state', ID: '$_id'}}"));
+  }
+
+  @Test public void testLimit() {
+    assertModel(MODEL)
+            .query("select state, id from zips\n"
+                    + "fetch next 3 rows only")
+            .runs()
+            .queryContains(
+                    mongoChecker(
+                            "{$limit: 3}",
+                            "{$project: {STATE: '$state', ID: '$_id'}}"));
+  }
+
+  @Ignore
+  @Test public void testFilterSort() {
+    // LONGITUDE and LATITUDE are null because of CALCITE-194.
+    Util.discard(Bug.CALCITE_194_FIXED);
+    assertModel(MODEL)
+            .query("select * from zips\n"
+                    + "where city = 'SPRINGFIELD' and id >= '70000'\n"
+                    + "order by state, id")
+            .returns(""
+                    + "CITY=SPRINGFIELD; LONGITUDE=null; LATITUDE=null; POP=752; STATE=AR; ID=72157\n"
+                    + "CITY=SPRINGFIELD; LONGITUDE=null; LATITUDE=null; POP=1992; STATE=CO; ID=81073\n"
+                    + "CITY=SPRINGFIELD; LONGITUDE=null; LATITUDE=null; POP=5597; STATE=LA; ID=70462\n"
+                    + "CITY=SPRINGFIELD; LONGITUDE=null; LATITUDE=null; POP=32384; STATE=OR; ID=97477\n"
+                    + "CITY=SPRINGFIELD; LONGITUDE=null; LATITUDE=null; POP=27521; STATE=OR; ID=97478\n")
+            .queryContains(
+                    mongoChecker(
+                            "{\n"
+                                    + "  $match: {\n"
+                                    + "    city: \"SPRINGFIELD\",\n"
+                                    + "    _id: {\n"
+                                    + "      $gte: \"70000\"\n"
+                                    + "    }\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {CITY: '$city', LONGITUDE: '$loc[0]', LATITUDE: '$loc[1]', POP: '$pop', STATE: '$state', ID: '$_id'}}",
+                            "{$sort: {STATE: 1, ID: 1}}"))
+            .explainContains("PLAN=MongoToEnumerableConverter\n"
+                    + "  MongoSort(sort0=[$4], sort1=[$5], dir0=[ASC], dir1=[ASC])\n"
+                    + "    MongoProject(CITY=[CAST(ITEM($0, 'city')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"], LONGITUDE=[CAST(ITEM(ITEM($0, 'loc'), 0)):FLOAT], LATITUDE=[CAST(ITEM(ITEM($0, 'loc'), 1)):FLOAT], POP=[CAST(ITEM($0, 'pop')):INTEGER], STATE=[CAST(ITEM($0, 'state')):VARCHAR(2) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"], ID=[CAST(ITEM($0, '_id')):VARCHAR(5) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"])\n"
+                    + "      MongoFilter(condition=[AND(=(CAST(ITEM($0, 'city')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\", 'SPRINGFIELD'), >=(CAST(ITEM($0, '_id')):VARCHAR(5) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\", '70000'))])\n"
+                    + "        MongoTableScan(table=[[mongo_raw, zips]])");
+  }
+
+  @Test public void testFilterSortDesc() {
+    assertModel(MODEL)
+            .query("select * from zips\n"
+                    + "where pop BETWEEN 45000 AND 46000\n"
+                    + "order by state desc, pop")
+            .limit(4)
+            .returns("CITY=BECKLEY; LONGITUDE=null; LATITUDE=null; POP=45196; STATE=WV; ID=25801\n"
+                    + "CITY=ROCKERVILLE; LONGITUDE=null; LATITUDE=null; POP=45328; STATE=SD; ID=57701\n"
+                    + "CITY=PAWTUCKET; LONGITUDE=null; LATITUDE=null; POP=45442; STATE=RI; ID=02860\n"
+                    + "CITY=LAWTON; LONGITUDE=null; LATITUDE=null; POP=45542; STATE=OK; ID=73505\n");
+  }
+
+  @Ignore("broken; [CALCITE-2115] is logged to fix it")
+  @Test public void testUnionPlan() {
+    assertModel(MODEL)
+            .query("select * from \"sales_fact_1997\"\n"
+                    + "union all\n"
+                    + "select * from \"sales_fact_1998\"")
+            .explainContains("PLAN=EnumerableUnion(all=[true])\n"
+                    + "  MongoToEnumerableConverter\n"
+                    + "    MongoProject(product_id=[CAST(ITEM($0, 'product_id')):DOUBLE])\n"
+                    + "      MongoTableScan(table=[[_foodmart, sales_fact_1997]])\n"
+                    + "  MongoToEnumerableConverter\n"
+                    + "    MongoProject(product_id=[CAST(ITEM($0, 'product_id')):DOUBLE])\n"
+                    + "      MongoTableScan(table=[[_foodmart, sales_fact_1998]])")
+            .limit(2)
+            .returns(
+                    MongoAssertions.checkResultUnordered(
+                            "product_id=337", "product_id=1512"));
+  }
+
+  @Ignore(
+          "java.lang.ClassCastException: java.lang.Integer cannot be cast to java.lang.Double")
+  @Test public void testFilterUnionPlan() {
+    assertModel(MODEL)
+            .query("select * from (\n"
+                    + "  select * from \"sales_fact_1997\"\n"
+                    + "  union all\n"
+                    + "  select * from \"sales_fact_1998\")\n"
+                    + "where \"product_id\" = 1")
+            .runs();
+  }
+
+  /** Tests that we don't generate multiple constraints on the same column.
+   * MongoDB doesn't like it. If there is an '=', it supersedes all other
+   * operators. */
+  @Test public void testFilterRedundant() {
+    assertModel(MODEL)
+            .query(
+                    "select * from zips where state > 'CA' and state < 'AZ' and state = 'OK'")
+            .runs()
+            .queryContains(
+                    mongoChecker(
+                            "{\n"
+                                    + "  \"$match\": {\n"
+                                    + "    \"state\": \"OK\"\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {CITY: '$city', LONGITUDE: '$loc[0]', LATITUDE: '$loc[1]', POP: '$pop', STATE: '$state', ID: '$_id'}}"));
+  }
+
+  @Test public void testSelectWhere() {
+    assertModel(MODEL)
+            .query(
+                    "select * from \"warehouse\" where \"warehouse_state_province\" = 'CA'")
+            .explainContains("PLAN=MongoToEnumerableConverter\n"
+                    + "  MongoProject(warehouse_id=[CAST(ITEM($0, 'warehouse_id')):DOUBLE], warehouse_state_province=[CAST(ITEM($0, 'warehouse_state_province')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"])\n"
+                    + "    MongoFilter(condition=[=(CAST(ITEM($0, 'warehouse_state_province')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\", 'CA')])\n"
+                    + "      MongoTableScan(table=[[mongo_raw, warehouse]])")
+            .returns(
+                    MongoAssertions.checkResultUnordered(
+                            "warehouse_id=6; warehouse_state_province=CA",
+                            "warehouse_id=7; warehouse_state_province=CA",
+                            "warehouse_id=14; warehouse_state_province=CA",
+                            "warehouse_id=24; warehouse_state_province=CA"))
+            .queryContains(
+                    // Per https://issues.apache.org/jira/browse/CALCITE-164,
+                    // $match must occur before $project for good performance.
+                    mongoChecker(
+                            "{\n"
+                                    + "  \"$match\": {\n"
+                                    + "    \"warehouse_state_province\": \"CA\"\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {warehouse_id: 1, warehouse_state_province: 1}}"));
+  }
+
+  @Test public void testInPlan() {
+    assertModel(MODEL)
+            .query("select \"store_id\", \"store_name\" from \"store\"\n"
+                    + "where \"store_name\" in ('Store 1', 'Store 10', 'Store 11', 'Store 15', 'Store 16', 'Store 24', 'Store 3', 'Store 7')")
+            .returns(
+                    MongoAssertions.checkResultUnordered(
+                            "store_id=1; store_name=Store 1",
+                            "store_id=3; store_name=Store 3",
+                            "store_id=7; store_name=Store 7",
+                            "store_id=10; store_name=Store 10",
+                            "store_id=11; store_name=Store 11",
+                            "store_id=15; store_name=Store 15",
+                            "store_id=16; store_name=Store 16",
+                            "store_id=24; store_name=Store 24"))
+            .queryContains(
+                    mongoChecker(
+                            "{\n"
+                                    + "  \"$match\": {\n"
+                                    + "    \"$or\": [\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 1\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 10\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 11\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 15\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 16\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 24\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 3\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"store_name\": \"Store 7\"\n"
+                                    + "      }\n"
+                                    + "    ]\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {store_id: 1, store_name: 1}}"));
+  }
+
+  /** Simple query based on the "mongo-zips" model. */
+  @Test public void testZips() {
+    assertModel(MODEL)
+            .query("select state, city from zips")
+            .returnsCount(ZIPS_SIZE);
+  }
+
+  @Test public void testCountGroupByEmpty() {
+    assertModel(MODEL)
+            .query("select count(*) from zips")
+            .returns(String.format(Locale.getDefault(), "EXPR$0=%d\n", ZIPS_SIZE))
+            .explainContains("PLAN=MongoToEnumerableConverter\n"
+                    + "  MongoAggregate(group=[{}], EXPR$0=[COUNT()])\n"
+                    + "    MongoTableScan(table=[[mongo_raw, zips]])")
+            .queryContains(
+                    mongoChecker(
+                            "{$group: {_id: {}, 'EXPR$0': {$sum: 1}}}"));
+  }
+
+  @Test public void testCountGroupByEmptyMultiplyBy2() {
+    // This operation is not supported by fongo: https://github.com/fakemongo/fongo/issues/152
+    MongoAssertions.assumeRealMongoInstance();
+
+    assertModel(MODEL)
+            .query("select count(*)*2 from zips")
+            .returns(String.format(Locale.getDefault(), "EXPR$0=%d\n", ZIPS_SIZE * 2))
+            .queryContains(
+                    mongoChecker(
+                            "{$group: {_id: {}, _0: {$sum: 1}}}",
+                            "{$project: {'EXPR$0': {$multiply: ['$_0', {$literal: 2}]}}}"));
+  }
+
+  @Test public void testGroupByOneColumnNotProjected() {
+    assertModel(MODEL)
+            .query("select count(*) from zips group by state order by 1")
+            .limit(2)
+            .returns("EXPR$0=2\n"
+                    + "EXPR$0=2\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', 'EXPR$0': {$sum: 1}}}",
+                            "{$project: {STATE: '$_id', 'EXPR$0': '$EXPR$0'}}",
+                            "{$project: {'EXPR$0': 1}}",
+                            "{$sort: {EXPR$0: 1}}"));
+  }
+
+  @Test public void testGroupByOneColumn() {
+    assertModel(MODEL)
+            .query(
+                    "select state, count(*) as c from zips group by state order by state")
+            .limit(2)
+            .returns("STATE=AK; C=3\nSTATE=AL; C=3\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', C: {$sum: 1}}}",
+                            "{$project: {STATE: '$_id', C: '$C'}}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Test public void testGroupByOneColumnReversed() {
+    // Note extra $project compared to testGroupByOneColumn.
+    assertModel(MODEL)
+            .query(
+                    "select count(*) as c, state from zips group by state order by state")
+            .limit(2)
+            .returns("C=3; STATE=AK\nC=3; STATE=AL\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', C: {$sum: 1}}}",
+                            "{$project: {STATE: '$_id', C: '$C'}}",
+                            "{$project: {C: 1, STATE: 1}}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Test public void testGroupByAvg() {
+    assertModel(MODEL)
+            .query(
+                    "select state, avg(pop) as a from zips group by state order by state")
+            .limit(2)
+            .returns("STATE=AK; A=26856\nSTATE=AL; A=43383\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {POP: '$pop', STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', A: {$avg: '$POP'}}}",
+                            "{$project: {STATE: '$_id', A: '$A'}}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Test public void testGroupByAvgSumCount() {
+    // This operation not supported by fongo: https://github.com/fakemongo/fongo/issues/152
+    MongoAssertions.assumeRealMongoInstance();
+    assertModel(MODEL)
+            .query(
+                    "select state, avg(pop) as a, sum(pop) as s, count(pop) as c from zips group by state order by state")
+            .limit(2)
+            .returns("STATE=AK; A=26856; S=80568; C=3\n"
+                    + "STATE=AL; A=43383; S=130151; C=3\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {POP: '$pop', STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', _1: {$sum: '$POP'}, _2: {$sum: {$cond: [ {$eq: ['POP', null]}, 0, 1]}}}}",
+                            "{$project: {STATE: '$_id', _1: '$_1', _2: '$_2'}}",
+                            "{$sort: {STATE: 1}}",
+                            "{$project: {STATE: 1, A: {$divide: [{$cond:[{$eq: ['$_2', {$literal: 0}]},null,'$_1']}, '$_2']}, S: {$cond:[{$eq: ['$_2', {$literal: 0}]},null,'$_1']}, C: '$_2'}}"));
+  }
+
+  @Test public void testGroupByHaving() {
+    assertModel(MODEL)
+            .query("select state, count(*) as c from zips\n"
+                    + "group by state having count(*) > 2 order by state")
+            .returnsCount(47)
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', C: {$sum: 1}}}",
+                            "{$project: {STATE: '$_id', C: '$C'}}",
+                            "{\n"
+                                    + "  \"$match\": {\n"
+                                    + "    \"C\": {\n"
+                                    + "      \"$gt\": 2\n"
+                                    + "    }\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Ignore("https://issues.apache.org/jira/browse/CALCITE-270")
+  @Test public void testGroupByHaving2() {
+    assertModel(MODEL)
+            .query("select state, count(*) as c from zips\n"
+                    + "group by state having sum(pop) > 12000000")
+            .returns("STATE=NY; C=1596\n"
+                    + "STATE=TX; C=1676\n"
+                    + "STATE=FL; C=826\n"
+                    + "STATE=CA; C=1523\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {STATE: '$state', POP: '$pop'}}",
+                            "{$group: {_id: '$STATE', C: {$sum: 1}, _2: {$sum: '$POP'}}}",
+                            "{$project: {STATE: '$_id', C: '$C', _2: '$_2'}}",
+                            "{\n"
+                                    + "  $match: {\n"
+                                    + "    _2: {\n"
+                                    + "      $gt: 12000000\n"
+                                    + "    }\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {STATE: 1, C: 1}}"));
+  }
+
+  @Test public void testGroupByMinMaxSum() {
+    assertModel(MODEL)
+            .query("select count(*) as c, state,\n"
+                    + " min(pop) as min_pop, max(pop) as max_pop, sum(pop) as sum_pop\n"
+                    + "from zips group by state order by state")
+            .limit(2)
+            .returns("C=3; STATE=AK; MIN_POP=23238; MAX_POP=32383; SUM_POP=80568\n"
+                    + "C=3; STATE=AL; MIN_POP=42124; MAX_POP=44165; SUM_POP=130151\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {POP: '$pop', STATE: '$state'}}",
+                            "{$group: {_id: '$STATE', C: {$sum: 1}, MIN_POP: {$min: '$POP'}, MAX_POP: {$max: '$POP'}, SUM_POP: {$sum: '$POP'}}}",
+                            "{$project: {STATE: '$_id', C: '$C', MIN_POP: '$MIN_POP', MAX_POP: '$MAX_POP', SUM_POP: '$SUM_POP'}}",
+                            "{$project: {C: 1, STATE: 1, MIN_POP: 1, MAX_POP: 1, SUM_POP: 1}}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Test public void testGroupComposite() {
+    assertModel(MODEL)
+            .query("select count(*) as c, state, city from zips\n"
+                    + "group by state, city\n"
+                    + "order by c desc, city\n"
+                    + "limit 2")
+            .returns("C=1; STATE=SD; CITY=ABERDEEN\n"
+                      + "C=1; STATE=SC; CITY=AIKEN\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {CITY: '$city', STATE: '$state'}}",
+                            "{$group: {_id: {CITY: '$CITY', STATE: '$STATE'}, C: {$sum: 1}}}",
+                            "{$project: {_id: 0, CITY: '$_id.CITY', STATE: '$_id.STATE', C: '$C'}}",
+                            "{$sort: {C: -1, CITY: 1}}",
+                            "{$limit: 2}",
+                            "{$project: {C: 1, STATE: 1, CITY: 1}}"));
+  }
+
+  @Ignore("broken; [CALCITE-2115] is logged to fix it")
+  @Test public void testDistinctCount() {
+    assertModel(MODEL)
+            .query("select state, count(distinct city) as cdc from zips\n"
+                    + "where state in ('CA', 'TX') group by state order by state")
+            .returns("STATE=CA; CDC=1072\n"
+                    + "STATE=TX; CDC=1233\n")
+            .queryContains(
+                    mongoChecker(
+                            "{\n"
+                                    + "  \"$match\": {\n"
+                                    + "    \"$or\": [\n"
+                                    + "      {\n"
+                                    + "        \"state\": \"CA\"\n"
+                                    + "      },\n"
+                                    + "      {\n"
+                                    + "        \"state\": \"TX\"\n"
+                                    + "      }\n"
+                                    + "    ]\n"
+                                    + "  }\n"
+                                    + "}",
+                            "{$project: {CITY: '$city', STATE: '$state'}}",
+                            "{$group: {_id: {CITY: '$CITY', STATE: '$STATE'}}}",
+                            "{$project: {_id: 0, CITY: '$_id.CITY', STATE: '$_id.STATE'}}",
+                            "{$group: {_id: '$STATE', CDC: {$sum: {$cond: [ {$eq: ['CITY', null]}, 0, 1]}}}}",
+                            "{$project: {STATE: '$_id', CDC: '$CDC'}}",
+                            "{$sort: {STATE: 1}}"));
+  }
+
+  @Test public void testDistinctCountOrderBy() {
+    // java.lang.ClassCastException: com.mongodb.BasicDBObject cannot be cast to java.lang.Number
+    // https://github.com/fakemongo/fongo/issues/152
+    MongoAssertions.assumeRealMongoInstance();
+    assertModel(MODEL)
+            .query("select state, count(distinct city) as cdc\n"
+                    + "from zips\n"
+                    + "group by state\n"
+                    + "order by cdc desc limit 5")
+            .returns("STATE=VA; CDC=3\n"
+                    + "STATE=NY; CDC=3\n"
+                    + "STATE=SC; CDC=3\n"
+                    + "STATE=RI; CDC=3\n"
+                    + "STATE=WV; CDC=3\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {CITY: '$city', STATE: '$state'}}",
+                            "{$group: {_id: {CITY: '$CITY', STATE: '$STATE'}}}",
+                            "{$project: {_id: 0, CITY: '$_id.CITY', STATE: '$_id.STATE'}}",
+                            "{$group: {_id: '$STATE', CDC: {$sum: {$cond: [ {$eq: ['CITY', null]}, 0, 1]}}}}",
+                            "{$project: {STATE: '$_id', CDC: '$CDC'}}",
+                            "{$sort: {CDC: -1}}",
+                            "{$limit: 5}"));
+  }
+
+  @Ignore("broken; [CALCITE-2115] is logged to fix it")
+  @Test public void testProject() {
+    assertModel(MODEL)
+            .query("select state, city, 0 as zero from zips order by state, city")
+            .limit(2)
+            .returns("STATE=AK; CITY=AKHIOK; ZERO=0\n"
+                    + "STATE=AK; CITY=AKIACHAK; ZERO=0\n")
+            .queryContains(
+                    mongoChecker(
+                            "{$project: {CITY: '$city', STATE: '$state'}}",
+                            "{$sort: {STATE: 1, CITY: 1}}",
+                            "{$project: {STATE: 1, CITY: 1, ZERO: {$literal: 0}}}"));
+  }
+
+  @Test public void testFilter() {
+    assertModel(MODEL)
+            .query("select state, city from zips where state = 'CA'")
+            .limit(2)
+            .returns("STATE=CA; CITY=LOS ANGELES\n"
+                    + "STATE=CA; CITY=BELL GARDENS\n")
+            .explainContains("PLAN=MongoToEnumerableConverter\n"
+                    + "  MongoProject(STATE=[CAST(ITEM($0, 'state')):VARCHAR(2) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"], CITY=[CAST(ITEM($0, 'city')):VARCHAR(20) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\"])\n"
+                    + "    MongoFilter(condition=[=(CAST(ITEM($0, 'state')):VARCHAR(2) CHARACTER SET \"ISO-8859-1\" COLLATE \"ISO-8859-1$en_US$primary\", 'CA')])\n"
+                    + "      MongoTableScan(table=[[mongo_raw, zips]])");
+  }
+
+  /** MongoDB's predicates are handed (they can only accept literals on the
+   * right-hand size) so it's worth testing that we handle them right both
+   * ways around. */
+  @Test public void testFilterReversed() {
+    assertModel(MODEL)
+            .query("select state, city from zips where 'WI' < state")
+            .limit(2)
+            .returns("STATE=WV; CITY=BECKLEY\nSTATE=WV; CITY=ELM GROVE\n");
+
+    assertModel(MODEL)
+            .query("select state, city from zips where state > 'WI'")
+            .limit(2)
+            .returns("STATE=WV; CITY=BECKLEY\n"
+                    + "STATE=WV; CITY=ELM GROVE\n");
+  }
+
+  /** MongoDB's predicates are handed (they can only accept literals on the
+   * right-hand size) so it's worth testing that we handle them right both
+   * ways around.
+   *
+   * <p>Test case for
+   * <a href="https://issues.apache.org/jira/browse/CALCITE-740">[CALCITE-740]
+   * Redundant WHERE clause causes wrong result in MongoDB adapter</a>. */
+  @Test public void testFilterPair() {
+    final int gt9k = 148;
+    final int lt9k = 1;
+    final int gt8k = 148;
+    final int lt8k = 1;
+    checkPredicate(gt9k, "where pop > 8000 and pop > 9000");
+    checkPredicate(gt9k, "where pop > 9000");
+    checkPredicate(lt9k, "where pop < 9000");
+    checkPredicate(gt8k, "where pop > 8000");
+    checkPredicate(lt8k, "where pop < 8000");
+    checkPredicate(gt9k, "where pop > 9000 and pop > 8000");
+    checkPredicate(gt8k, "where pop > 9000 or pop > 8000");
+    checkPredicate(gt8k, "where pop > 8000 or pop > 9000");
+    checkPredicate(lt8k, "where pop < 8000 and pop < 9000");
+  }
+
+  private void checkPredicate(int expected, String q) {
+    assertModel(MODEL)
+            .query("select count(*) as c from zips\n"
+                    + q)
+            .returns("C=" + expected + "\n");
+    assertModel(MODEL)
+            .query("select * from zips\n"
+                    + q)
+            .returnsCount(expected);
+  }
+
+  /** Test case for
+   * <a href="https://issues.apache.org/jira/browse/CALCITE-286">[CALCITE-286]
+   * Error casting MongoDB date</a>. */
+  @Test public void testDate() {
+    // Assumes that you have created the following collection before running
+    // this test:
+    //
+    // $ mongo
+    // > use test
+    // switched to db test
+    // > db.createCollection("datatypes")
+    // { "ok" : 1 }
+    // > db.datatypes.insert( {
+    //     "_id" : ObjectId("53655599e4b0c980df0a8c27"),
+    //     "_class" : "com.ericblue.Test",
+    //     "date" : ISODate("2012-09-05T07:00:00Z"),
+    //     "value" : 1231,
+    //     "ownerId" : "531e7789e4b0853ddb861313"
+    //   } )
+    assertModel("{\n"
+            + "  version: '1.0',\n"
+            + "  defaultSchema: 'test',\n"
+            + "   schemas: [\n"
+            + "     {\n"
+            + "       type: 'custom',\n"
+            + "       name: 'test',\n"
+            + "       factory: 'org.apache.calcite.adapter.mongodb.MongoSchemaFactory',\n"
+            + "       operand: {\n"
+            + "         host: 'localhost',\n"
+            + "         database: 'test'\n"
+            + "       }\n"
+            + "     }\n"
+            + "   ]\n"
+            + "}")
+            .query("select cast(_MAP['date'] as DATE) from \"datatypes\"")
+            .returnsUnordered("EXPR$0=2012-09-05");
+  }
+
+  /** Test case for
+   * <a href="https://issues.apache.org/jira/browse/CALCITE-665">[CALCITE-665]
+   * ClassCastException in MongoDB adapter</a>. */
+  @Test public void testCountViaInt() {
+    assertModel(MODEL)
+        .query("select count(*) from zips")
+        .returns(
+            new Function<ResultSet, Void>() {
+              public Void apply(ResultSet input) {
+                try {
+                  Assert.assertThat(input.next(), CoreMatchers.is(true));
+                  Assert.assertThat(input.getInt(1), CoreMatchers.is(ZIPS_SIZE));
+                  return null;
+                } catch (SQLException e) {
+                  throw new RuntimeException(e);
+                }
+              }
+            });
+  }
+
+  /** Returns a function that checks that a particular MongoDB pipeline is
+   * generated to implement a query. */
+  private static Function<List, Void> mongoChecker(final String... strings) {
+    return new Function<List, Void>() {
+      public Void apply(List actual) {
+        Object[] actualArray =
+                actual == null || actual.isEmpty()
+                        ? null
+                        : ((List) actual.get(0)).toArray();
+        CalciteAssert.assertArrayEqual("expected MongoDB query not found",
+                strings, actualArray);
+        return null;
       }
     };
-  }
-
-  @Test
-  public void single() {
-    zips.insertOne(new Document());
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select * from \"mongo\".\"zips\"")
-            .runs()
-            .returnsCount(1);
-  }
-
-  @Test
-  public void empty() {
-    // for some reason fongo doesn't list collection if it was unused
-    zips.insertOne(new Document());
-    zips.deleteMany(new BsonDocument());
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select * from \"mongo\".\"zips\"")
-            .runs()
-            .returnsCount(0);
-  }
-
-  @Test
-  public void filter() {
-    zips.insertOne(new Document("CITY", "New York").append("STATE", "NY"));
-    zips.insertOne(new Document("CITY", "Washington").append("STATE", "DC"));
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select cast(_MAP['CITY'] as varchar(20)) as \"city\" from \"mongo\".\"zips\" "
-                    + " where _MAP['STATE'] = 'NY'")
-            .returns("city=New York\n");
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select cast(_MAP['CITY'] as varchar(20)) as \"city\" from \"mongo\".\"zips\" "
-                   + " where _MAP['STATE'] = 'DC'")
-            .returns("city=Washington\n");
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select cast(_MAP['CITY'] as varchar(20)) as \"city\" from \"mongo\".\"zips\" "
-                    + " where _MAP['STATE'] in ('DC', 'NY')")
-            .returns("city=New York\ncity=Washington\n");
-  }
-
-  @Test
-  public void limit() {
-    zips.insertOne(new Document("CITY", "New York").append("STATE", "NY"));
-    zips.insertOne(new Document("CITY", "Washington").append("STATE", "DC"));
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select * from \"mongo\".\"zips\" limit 1")
-            .returnsCount(1);
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select * from \"mongo\".\"zips\" limit 2")
-            .returnsCount(2);
-
-  }
-
-  /**
-   * Following queries are not supported in Mongo adapter :
-   * <pre>
-   * {@code A and (B or C)}
-   * {@code (A or B) and C}
-   * </pre>
-   *
-   * @see <a href="https://issues.apache.org/jira/browse/CALCITE-2331">[CALCITE-2331]</a>
-   */
-  @Ignore("broken; [CALCITE-2331] is logged to fix it")
-  @Test
-  public void validateCALCITE2331() {
-    zips.insertOne(new Document("CITY", "New York").append("STATE", "NY"));
-    zips.insertOne(new Document("CITY", "Washington").append("STATE", "DC"));
-
-    CalciteAssert.that()
-            .with(newConnectionFactory())
-            .query("select cast(_MAP['CITY'] as varchar(20)) as \"city\" from \"mongo\".\"zips\" "
-                    + " where _MAP['STATE'] in ('DC', 'NY') and _MAP['CITY'] = 'New York'")
-            .returns("city=New York\n");
   }
 }
