+++ mongodb/src/test/resources/mongo-model.json
@@ -28,12 +28,41 @@
       }
     },
     {
+      "type": "custom",
+      "name": "_foodmart",
+      "factory": "org.apache.calcite.adapter.mongodb.MongoSchemaFactory",
+      "operand": {
+        "host": "localhost",
+        "database": "foodmart"
+      }
+    },
+    {
       "name": "mongo",
       "tables": [
         {
           "name": "ZIPS",
           "type": "view",
           "sql": "select cast(_MAP['city'] AS varchar(20)) AS city,\n cast(_MAP['loc'][0] AS float) AS longitude, cast(_MAP['loc'][1] AS float) AS latitude, cast(_MAP['pop'] AS integer) AS pop, cast(_MAP['state'] AS varchar(2)) AS state, cast(_MAP['_id'] AS varchar(5)) AS id from \"mongo_raw\".\"zips\""
+        },
+        {
+          "name": "sales_fact_1997",
+          "type": "view",
+          "sql": "select cast(_MAP['product_id'] AS double) AS \"product_id\" from \"_foodmart\".\"sales_fact_1997\""
+        },
+        {
+          "name": "sales_fact_1998",
+          "type": "view",
+          "sql": "select cast(_MAP['product_id'] AS double) AS \"product_id\" from \"_foodmart\".\"sales_fact_1998\""
+        },
+        {
+          "name": "store",
+          "type": "view",
+          "sql": "select cast(_MAP['store_id'] AS double) AS \"store_id\", cast(_MAP['store_name'] AS varchar(20)) AS \"store_name\" from \"mongo_raw\".\"store\""
+        },
+        {
+          "name": "warehouse",
+          "type": "view",
+          "sql": "select cast(_MAP['warehouse_id'] AS double) AS \"warehouse_id\", cast(_MAP['warehouse_state_province'] AS varchar(20)) AS \"warehouse_state_province\" from \"mongo_raw\".\"warehouse\""
         }
       ]
     }
