+++ pom.xml
@@ -78,6 +78,7 @@ limitations under the License.
     <git-commit-id-plugin.version>2.1.9</git-commit-id-plugin.version>
     <geode.version>1.3.0</geode.version>
     <fongo.version>2.1.1</fongo.version>
+    <foodmart-data-json.version>0.4</foodmart-data-json.version>
 
     <!-- We support (and test against) Guava versions between
          19.0 and 23.0. Default is 19.0 due to Cassandra adapter. -->
@@ -465,6 +466,11 @@ limitations under the License.
         <version>${fongo.version}</version>
       </dependency>
       <dependency>
+        <groupId>net.hydromatic</groupId>
+        <artifactId>foodmart-data-json</artifactId>
+        <version>${foodmart-data-json.version}</version>
+      </dependency>
+      <dependency>
         <groupId>org.openjdk.jmh</groupId>
         <artifactId>jmh-core</artifactId>
         <version>${jmh.version}</version>
@@ -828,6 +834,9 @@ limitations under the License.
                 <threadCount>6</threadCount>
                 <parallel>both</parallel>
                 <argLine>-Xmx1024m</argLine>
+                <systemPropertyVariables>
+                  <calcite.integrationTest>true</calcite.integrationTest>
+                </systemPropertyVariables>
               </configuration>
             </execution>
             <execution>
